Update command-pattern-with-autowirelocator.md - Round #1
I changed a lot of wording throughout to improve flow or account for things happening on screen that wasn't really being addressed by the text. I also added some code to the script because I feel like it's important, especially for the sake of not having big silent gaps, to talk through what we're typing a bit more.

And this is just a note for the video portion: There was A LOT of movement on screen that isn't super necessary. Mouse movements in big loops, a lot of pointing to things that were ultimately removed or simplified in the script, scrolling up and down kind of quick when there was no need, etc.
Just something to keep in mind for future recordings.